+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 # Changelog
 
 =====================
+## [0.2.0] 20/11/2022 - `Documentation & Scripts`: Docs & scripts updates and many more
+
+Changes:
+
+0. VERSION and CHANGELOG created
+1. README.md updated
+2. docker-compose.yaml fix
+3. `latest` tag added
+4. Management scripts added
+5. OpenVPN version `2.5.8` added
+
+=====================
 ## [0.1.0] 30/10/2022 - `Basic stuff`: Bare minimum
 
 Changes: